Scene Graph, Tabletop and Pickagram Hierarchy, Camera Left Mouse Button Controls
Topics Done (Geometry):

To manage your scene, use a simple scene graph in which each node has a pointer to the
mesh to be drawn, a program shader to be active while the geometry is being drawn, and
the ModelMatrix to be used for positioning the model. Implement a hierarchical
composition in your scene graph so that each node takes its parent’s configuration into
account when it is drawn

The Pickagram must be placed on top of a planar surface, which will act as a tabletop, for
instance. Therefore, if you move the surface in the scene, the Pickagram on it should
move along with it. The Pickagram can also move independently of the surface (see the
Animation section).

To move the surface, press and hold down the left mouse button while dragging the mouse. If you drag the mouse to the left, it moves along the left‑side vector of the
camera’s direction. If you drag the mouse to the right, it moves along the right‑side vector. of the camera’s direction. If you drag the mouse upwards, the surface (and all above)
should move in the direction of the front vector of the camera, away from the camera. If you drag the mouse downwards, the surface should move towards the camera.